Added base service class
